Revert "Gbe 400 Fast filter (#402)"
This reverts commit b3297d35c2427d2555f14e0b4de94e4ccf1c78af.
