+++ gbe/templates/gbe/email/filter_checkbox_horizontal.tmpl
