+++ gbe/templates/gbe/email/mail_to_group.tmpl
@@ -58,7 +58,7 @@
           </div>
           <div id="to_list" class="panel-collapse collapse">
             <div class="card-body">
-              {% include "filter_checkbox_horizontal.tmpl" with field=email_form.to %}
+              {% include "gbe/email/filter_checkbox_horizontal.tmpl" with field=email_form.to %}
             </div>
           </div>
       </div><!-- end panel group -->
