+++ gbe/templates/gbe/email/mail_to_roles.tmpl
@@ -9,24 +9,24 @@
   <form class="form-inline" action="" method="post" enctype="multipart/form-data">
     {% csrf_token %}
     <div class="row email-select py-2">
-    {% include "filter_checkbox_horizontal.tmpl" with field=selection_form.conference %}
+    {% include "gbe/email/filter_checkbox_horizontal.tmpl" with field=selection_form.conference %}
     </div>
     <div class="row email-select py-2">
-    {% include "filter_checkbox_horizontal.tmpl" with field=selection_form.roles %}
+    {% include "gbe/email/filter_checkbox_horizontal.tmpl" with field=selection_form.roles %}
     {% if specify_event_form.events in specify_event_form.visible_fields %}
     </div>
     <div class="row email-select py-2">
-      {% include "filter_checkbox_horizontal.tmpl" with field=specify_event_form.events %}
+      {% include "gbe/email/filter_checkbox_horizontal.tmpl" with field=specify_event_form.events %}
     {% endif%}
     {% if specify_event_form.staff_areas in specify_event_form.visible_fields %}
     </div>
     <div class="row email-select py-2">
-      {% include "filter_checkbox_horizontal.tmpl" with field=specify_event_form.staff_areas %}
+      {% include "gbe/email/filter_checkbox_horizontal.tmpl" with field=specify_event_form.staff_areas %}
     {% endif%}
     {% if specify_event_form.event_collections in specify_event_form.visible_fields %}
     </div>
     <div class="row email-select py-2">
-      {% include "filter_checkbox_horizontal.tmpl" with field=specify_event_form.event_collections %}
+      {% include "gbe/email/filter_checkbox_horizontal.tmpl" with field=specify_event_form.event_collections %}
     {% endif%}
       <div class="form-group col-md-1 col-12">
         {% if specify_event_form %}
