+++ gbe/templates/gbe/gbe_table.tmpl
@@ -24,7 +24,7 @@
     {% csrf_token %}
     {% for field in filter_form.visible_fields %}
     <div class="row email-select">
-      {% include "filter_checkbox_horizontal.tmpl" %}
+      {% include "gbe/email/filter_checkbox_horizontal.tmpl" %}
     </div>
     <div>
     <input type="submit" class="btn gbe-btn-secondary float-right" name="filter" value="Filter Interest">
