+++ templates/conference_list.tmpl
@@ -2,5 +2,5 @@
      {% if conf_slug == slug %}<font style="font-weight: bold">
      {% else %}<font style="font-weight: normal">{% endif %}
      <a href = "?conference={{ slug }}" class="gbe-link">
-     {{slug}}</a></font>{% if not forloop.last %}&nbsp;|&nbsp;{% endif %}
+     {{slug}}</a></font>&nbsp;|&nbsp;
    {% endfor %}
